# ingester: ActiveSeriesRequest can return native histograms information