--- a/pkg/ingester/active_series.go
+++ b/pkg/ingester/active_series.go
@@ -9,7 +9,9 @@ import (
 	"github.com/go-kit/log/level"
 	"github.com/grafana/dskit/tenant"
 	"github.com/prometheus/prometheus/model/labels"
+	"github.com/prometheus/prometheus/storage"
 	"github.com/prometheus/prometheus/tsdb"
+	"github.com/prometheus/prometheus/tsdb/index"
 
 	"github.com/grafana/mimir/pkg/ingester/activeseries"
 	"github.com/grafana/mimir/pkg/ingester/client"
@@ -56,15 +58,30 @@ func (i *Ingester) ActiveSeries(request *client.ActiveSeriesRequest, stream clie
 		return nil
 	}
 
-	series, err := listActiveSeries(ctx, db, matchers)
+	idx, err := db.Head().Index()
+	if err != nil {
+		return fmt.Errorf("error getting index: %w", err)
+	}
+
+	isNativeHistogram := request.GetType() == client.NATIVE_HISTOGRAM_SERIES
+	postings, err := getPostings(ctx, db, idx, matchers, isNativeHistogram)
 	if err != nil {
 		return fmt.Errorf("error listing active series: %w", err)
 	}
 
+	buf := labels.NewScratchBuilder(10)
 	resp := &client.ActiveSeriesResponse{}
 	currentSize := 0
-	for series.Next() {
-		m := &mimirpb.Metric{Labels: mimirpb.FromLabelsToLabelAdapters(series.At())}
+	for postings.Next() {
+		seriesRef, count := postings.AtBucketCount()
+		err = idx.Series(seriesRef, &buf, nil)
+		if err != nil {
+			return fmt.Errorf("error getting series: %w", err)
+		}
+		if isNativeHistogram {
+			buf.Add("__nh_bucket_count__", fmt.Sprintf("%d", count))
+		}
+		m := &mimirpb.Metric{Labels: mimirpb.FromLabelsToLabelAdapters(buf.Labels())}
 		mSize := m.Size()
 		if currentSize+mSize > activeSeriesMaxSizeBytes {
 			if err := client.SendActiveSeriesResponse(stream, resp); err != nil {
@@ -76,7 +93,7 @@ func (i *Ingester) ActiveSeries(request *client.ActiveSeriesRequest, stream clie
 		resp.Metric = append(resp.Metric, m)
 		currentSize += mSize
 	}
-	if err := series.Err(); err != nil {
+	if err := postings.Err(); err != nil {
 		return fmt.Errorf("error iterating over series: %w", err)
 	}
 
@@ -89,13 +106,7 @@ func (i *Ingester) ActiveSeries(request *client.ActiveSeriesRequest, stream clie
 	return nil
 }
 
-// listActiveSeries returns an iterator over the active series matching the given matchers.
-func listActiveSeries(ctx context.Context, db *userTSDB, matchers []*labels.Matcher) (series *Series, err error) {
-	idx, err := db.Head().Index()
-	if err != nil {
-		return nil, fmt.Errorf("error getting index: %w", err)
-	}
-
+func getPostings(ctx context.Context, db *userTSDB, idx tsdb.IndexReader, matchers []*labels.Matcher, isNativeHistogram bool) (activeseries.BucketCountPostings, error) {
 	if db.activeSeries == nil {
 		return nil, fmt.Errorf("active series tracker is not initialized")
 	}
@@ -110,11 +121,38 @@ func listActiveSeries(ctx context.Context, db *userTSDB, matchers []*labels.Matc
 		return nil, fmt.Errorf("error getting postings: %w", err)
 	}
 
-	postings = activeseries.NewPostings(db.activeSeries, postings)
-
 	if shard != nil {
 		postings = idx.ShardedPostings(postings, shard.ShardIndex, shard.ShardCount)
 	}
 
+	if isNativeHistogram {
+		return activeseries.NewNativeHistogramPostings(db.activeSeries, postings), nil
+	}
+
+	return &ZeroBucketCountPostings{*activeseries.NewPostings(db.activeSeries, postings)}, nil
+}
+
+// listActiveSeries is used for testing purposes, builds the whole array of active series in memory.
+func listActiveSeries(ctx context.Context, db *userTSDB, matchers []*labels.Matcher) (series *Series, err error) {
+	idx, err := db.Head().Index()
+	if err != nil {
+		return nil, fmt.Errorf("error getting index: %w", err)
+	}
+	postings, err := getPostings(ctx, db, idx, matchers, false)
+	if err != nil {
+		return nil, err
+	}
 	return NewSeries(postings, idx), nil
 }
+
+type ZeroBucketCountPostings struct {
+	activeseries.Postings
+}
+
+func (z *ZeroBucketCountPostings) AtBucketCount() (storage.SeriesRef, int) {
+	return z.At(), 0
+}
+
+// Type check.
+var _ index.Postings = &ZeroBucketCountPostings{}
+var _ activeseries.BucketCountPostings = &ZeroBucketCountPostings{}