--- a/pkg/ingester/active_series_test.go
+++ b/pkg/ingester/active_series_test.go
@@ -15,6 +15,8 @@ import (
 
 	"github.com/grafana/mimir/pkg/ingester/client"
 	"github.com/grafana/mimir/pkg/mimirpb"
+	util_test "github.com/grafana/mimir/pkg/util/test"
+	"github.com/grafana/mimir/pkg/util/validation"
 )
 
 func TestIngester_ActiveSeries(t *testing.T) {
@@ -46,7 +48,7 @@ func TestIngester_ActiveSeries(t *testing.T) {
 	}
 
 	// Write the series.
-	ingesterClient := prepareHealthyIngester(t)
+	ingesterClient := prepareHealthyIngester(t, nil)
 	ctx := user.InjectOrgID(context.Background(), userID)
 	_, err := ingesterClient.Push(ctx, writeReq)
 	require.NoError(t, err)
@@ -65,8 +67,79 @@ func TestIngester_ActiveSeries(t *testing.T) {
 	returnedSeriesCount := 0
 	for _, res := range server.responses {
 		returnedSeriesCount += len(res.Metric)
+		// Check that all series has the expected number of labels.
+		for _, m := range res.Metric {
+			assert.Equal(t, 2, len(m.Labels))
+		}
 	}
 	assert.Equal(t, len(writeReq.Timeseries), returnedSeriesCount)
+
+	// Check that we got the correct number of messages.
+	assert.Equal(t, expectedMessageCount, len(server.responses))
+}
+
+func TestIngester_ActiveNativeHistogramSeries(t *testing.T) {
+	samples := []mimirpb.Sample{{TimestampMs: 1_000, Value: 1}}
+	histograms := []mimirpb.Histogram{mimirpb.FromHistogramToHistogramProto(1_000, util_test.GenerateTestHistogram(1))}
+
+	seriesWithLabelsOfSize := func(size, index int, isHistogram bool) mimirpb.PreallocTimeseries {
+		// 24 bytes of static strings and slice overhead, the remaining bytes are used to
+		// pad the value of the "lbl" label.
+		require.Greater(t, size, 24, "minimum message size is 24 bytes")
+		tpl := fmt.Sprintf("%%0%dd", size-24)
+		ts := &mimirpb.TimeSeries{
+			Labels: mimirpb.FromLabelsToLabelAdapters(labels.FromStrings(labels.MetricName, "test", "lbl", fmt.Sprintf(tpl, index))),
+		}
+		if isHistogram {
+			ts.Histograms = histograms
+		} else {
+			ts.Samples = samples
+		}
+		return mimirpb.PreallocTimeseries{TimeSeries: ts}
+	}
+
+	expectedMessageCount := 4
+	totalSeriesSize := expectedMessageCount * activeSeriesMaxSizeBytes
+
+	writeReq := &mimirpb.WriteRequest{Source: mimirpb.API}
+	currentSize := 0
+	for i := 0; currentSize < totalSeriesSize; i++ {
+		isHistogram := i%2 != 0 // Half of the series will be float and the other half will be native histograms.
+		s := seriesWithLabelsOfSize(1024, i, isHistogram)
+		writeReq.Timeseries = append(writeReq.Timeseries, s)
+		if isHistogram {
+			currentSize += s.Size()
+		}
+	}
+
+	// Write the series.
+	ingesterClient := prepareHealthyIngester(t, func(limits *validation.Limits) { limits.NativeHistogramsIngestionEnabled = true })
+	ctx := user.InjectOrgID(context.Background(), userID)
+	_, err := ingesterClient.Push(ctx, writeReq)
+	require.NoError(t, err)
+
+	// Get active series
+	req, err := client.ToActiveSeriesRequest([]*labels.Matcher{
+		labels.MustNewMatcher(labels.MatchEqual, labels.MetricName, "test"),
+	})
+	req.Type = client.NATIVE_HISTOGRAM_SERIES
+	require.NoError(t, err)
+
+	server := &mockActiveSeriesServer{ctx: ctx}
+	err = ingesterClient.ActiveSeries(req, server)
+	require.NoError(t, err)
+
+	// Check that all series were returned.
+	returnedSeriesCount := 0
+	for _, res := range server.responses {
+		returnedSeriesCount += len(res.Metric)
+		// Check that all series contain the special label "__nh_bucket_count__".
+		for _, m := range res.Metric {
+			assert.Equal(t, 3, len(m.Labels))
+			assert.Equal(t, "8", mimirpb.FromLabelAdaptersToLabels(m.Labels).Get("__nh_bucket_count__"))
+		}
+	}
+	assert.Equal(t, len(writeReq.Timeseries)/2, returnedSeriesCount)
 
 	// Check that we got the correct number of messages.
 	assert.Equal(t, expectedMessageCount, len(server.responses))
@@ -79,7 +152,7 @@ func BenchmarkIngester_ActiveSeries(b *testing.B) {
 		metricName = "metric_name"
 	)
 
-	in := prepareHealthyIngester(b)
+	in := prepareHealthyIngester(b, nil)
 	ctx := user.InjectOrgID(context.Background(), userID)
 
 	samples := []mimirpb.Sample{{TimestampMs: 1_000, Value: 1}}