--- a/pkg/ingester/ingester_test.go
+++ b/pkg/ingester/ingester_test.go
@@ -5282,11 +5282,14 @@ func mockWriteRequest(t testing.TB, lbls labels.Labels, value float64, timestamp
 	return req, expectedQueryRes, expectedQueryStreamResSamples, expectedQueryStreamResChunks
 }
 
-func prepareHealthyIngester(b testing.TB) *Ingester {
+func prepareHealthyIngester(b testing.TB, mutateLimits func(*validation.Limits)) *Ingester {
 	cfg := defaultIngesterTestConfig(b)
 	limits := defaultLimitsTestConfig()
 	limits.MaxGlobalSeriesPerMetric = 0
 	limits.MaxGlobalSeriesPerUser = 0
+	if mutateLimits != nil {
+		mutateLimits(&limits)
+	}
 
 	// Create ingester.
 	i, err := prepareIngesterWithBlocksStorageAndLimits(b, cfg, limits, nil, "", nil)