@krajorama 2024-04-30
```
$ benchstat old.txt new.txt 
goos: linux
goarch: amd64
pkg: github.com/grafana/mimir/pkg/ingester
cpu: AMD Ryzen 7 4700U with Radeon Graphics         
                                       │   old.txt   │              new.txt               │
                                       │   sec/op    │   sec/op     vs base               │
Ingester_ActiveSeries/few_series-8       4.848µ ± 2%   4.761µ ± 1%  -1.77% (p=0.025 n=10)
Ingester_ActiveSeries/~10%_of_series-8   2.220m ± 3%   2.122m ± 2%  -4.40% (p=0.000 n=10)
geomean                                  103.7µ        100.5µ       -3.10%

                                       │   old.txt    │               new.txt               │
                                       │     B/op     │     B/op      vs base               │
Ingester_ActiveSeries/few_series-8       2.461Ki ± 0%   2.406Ki ± 0%  -2.22% (p=0.000 n=10)
Ingester_ActiveSeries/~10%_of_series-8   618.4Ki ± 0%   618.3Ki ± 0%  -0.01% (p=0.000 n=10)
geomean                                  39.01Ki        38.57Ki       -1.12%

                                       │   old.txt   │               new.txt                │
                                       │  allocs/op  │  allocs/op   vs base                 │
Ingester_ActiveSeries/few_series-8        48.00 ± 0%    48.00 ± 0%       ~ (p=1.000 n=10) ¹
Ingester_ActiveSeries/~10%_of_series-8   8.437k ± 0%   8.437k ± 0%       ~ (p=1.000 n=10) ¹
geomean                                   636.4         636.4       +0.00%
¹ all samples are equal
```