@krajorama 2024-05-01
```
$ benchstat old.txt new.txt 
goos: linux
goarch: amd64
pkg: github.com/grafana/mimir/pkg/ingester
cpu: AMD Ryzen 7 4700U with Radeon Graphics         
                                       │   old.txt    │               new.txt               │
                                       │    sec/op    │   sec/op     vs base                │
Ingester_ActiveSeries/few_series-8       4.824µ ±  6%   4.727µ ± 1%   -2.01% (p=0.007 n=10)
Ingester_ActiveSeries/~10%_of_series-8   2.506m ± 12%   2.222m ± 1%  -11.36% (p=0.005 n=10)
geomean                                  110.0µ         102.5µ        -6.80%

                                       │   old.txt    │               new.txt               │
                                       │     B/op     │     B/op      vs base               │
Ingester_ActiveSeries/few_series-8       2.461Ki ± 0%   2.430Ki ± 0%  -1.27% (p=0.000 n=10)
Ingester_ActiveSeries/~10%_of_series-8   618.4Ki ± 0%   618.3Ki ± 0%  -0.01% (p=0.000 n=10)
geomean                                  39.01Ki        38.76Ki       -0.64%

                                       │   old.txt   │               new.txt                │
                                       │  allocs/op  │  allocs/op   vs base                 │
Ingester_ActiveSeries/few_series-8        48.00 ± 0%    48.00 ± 0%       ~ (p=1.000 n=10) ¹
Ingester_ActiveSeries/~10%_of_series-8   8.437k ± 0%   8.437k ± 0%       ~ (p=1.000 n=10) ¹
geomean                                   636.4         636.4       +0.00%
¹ all samples are equal
```